+++ src/types/cellType.ts
@@ -1,7 +1,8 @@
-export type CellData = {
-  id?: string;
+export type CellType = {
+  id?: string; // cell의 고유 ID
   value?: string;
-  isMainCenter?: boolean;
-  isCenter?: boolean;
   placeholder?: string;
+  isMainCenter?: boolean; // 중앙 3x3 grid의 여부
+  isCenter?: boolean; // 각 블록의 중앙 칸 여부
+  onChange: (val: string) => void;
 };
